+++ front-end/src/app/app.component.spec.ts
@@ -1,6 +1,10 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterModule } from '@angular/router';
+import { provideMockStore } from '@ngrx/store/testing';
 import { AppComponent } from './app.component';
+import { LoginService } from './shared/services/login.service';
+import { testMockStore } from './shared/utils/unit-test.utils';
 
 describe('AppComponent', () => {
   let component: AppComponent;
@@ -8,7 +12,8 @@ describe('AppComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [RouterModule.forRoot([])],
+      imports: [HttpClientTestingModule, RouterModule.forRoot([])],
+      providers: [LoginService, provideMockStore(testMockStore)],
       declarations: [AppComponent],
     }).compileComponents();
   });
