+++ front-end/src/app/app.component.ts
@@ -1,12 +1,16 @@
-import { Component } from '@angular/core';
-import { DestroyerComponent } from "./shared/components/app-destroyer.component";
+import { Component, OnInit } from '@angular/core';
+import { LoginService } from './shared/services/login.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
 })
-export class AppComponent extends DestroyerComponent {
+export class AppComponent implements OnInit {
+  constructor(private loginService: LoginService) { }
 
+  ngOnInit(): void {
+    this.loginService.dispatchUserLoggedInFromCookies();
+  }
 }
