+++ front-end/src/app/dashboard/dashboard.component.spec.ts
@@ -2,12 +2,6 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
-import { Store } from '@ngrx/store';
-import { provideMockStore } from '@ngrx/store/testing';
-import { UserLoginData } from 'app/shared/models/user.model';
-import { testMockStore } from 'app/shared/utils/unit-test.utils';
-import { userLoggedInAction } from 'app/store/login.actions';
-import { environment } from 'environments/environment';
 import { CookieService } from 'ngx-cookie-service';
 import { PanelModule } from 'primeng/panel';
 import { DashboardComponent } from './dashboard.component';
@@ -16,13 +10,12 @@ describe('DashboardComponent', () => {
   let component: DashboardComponent;
   let fixture: ComponentFixture<DashboardComponent>;
   let cookieService: CookieService;
-  let store: Store;
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [HttpClientTestingModule, RouterTestingModule, PanelModule, BrowserAnimationsModule],
       declarations: [DashboardComponent],
-      providers: [CookieService, provideMockStore(testMockStore)],
+      providers: [],
     }).compileComponents();
   }));
 
@@ -31,44 +24,10 @@ describe('DashboardComponent', () => {
     component = fixture.componentInstance;
     fixture.detectChanges();
     cookieService = TestBed.inject(CookieService);
-    store = TestBed.inject(Store);
   });
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
 
-  it('#dispatchUserLoggedInFromCookies happy path', () => {
-    const testFirstName = 'testFirstName';
-    const testLastName = 'testLastName';
-    const testEmail = 'testEmail';
-    const testLoginDotGov = false;
-
-    const expectedUserLoginData: UserLoginData = {
-      first_name: testFirstName,
-      last_name: testLastName,
-      email: testEmail,
-      login_dot_gov: testLoginDotGov
-    };
-    spyOn(cookieService, 'check').and.returnValue(true);
-    spyOn(cookieService, 'get').and.callFake((name: string) => {
-      if (name === environment.ffapiFirstNameCookieName) {
-        return testFirstName;
-      }
-      if (name === environment.ffapiLastNameCookieName) {
-        return testLastName;
-      }
-      if (name === environment.ffapiEmailCookieName) {
-        return testEmail;
-      }
-      if (name === environment.ffapiLoginDotGovCookieName) {
-        return testLoginDotGov.toString();
-      }
-      throw Error('fail!');
-    });
-    spyOn(store, 'dispatch');
-
-    component.dispatchUserLoggedInFromCookies();
-    expect(store.dispatch).toHaveBeenCalledWith(userLoggedInAction({ payload: expectedUserLoginData }));
-  });
 });
